requirements.txt
# todo: add all dependencies
numpy
PyYAML==5.3.1
self==0.0.0
serial==0.0.97
